--- list.c
@@ -1,1 +1,0 @@
-#include "list.h"
