+++ movegen.c
@@ -1,7 +1,7 @@
 #include "movegen.h"
 #include <string.h>
 #include "bitboard.h"
-#include "list.h"
+#include "movelist.h"
 #include "position.h"
 
 void generate_pawn_moves(const GameState *pos, int turn, int piece_offset, MoveList *move_list)
@@ -28,7 +28,7 @@ void generate_pawn_moves(const GameState *pos, int turn, int piece_offset, MoveL
     while (double_push_target) {
         dst = GET_FIRST_BIT_SQUARE(double_push_target);
         src = dst - 16 + (32 * turn);
-        move_list->list[i++] = CREATE_MOVE(src, dst, piece, NO_PROMOTION, NO_CAPTURE, 1, 0, 0);
+        move_list->move[i++] = CREATE_MOVE(src, dst, piece, NO_PROMOTION, NO_CAPTURE, 1, 0, 0);
         CLEAR_LSB(double_push_target);
     }
 
@@ -36,13 +36,13 @@ void generate_pawn_moves(const GameState *pos, int turn, int piece_offset, MoveL
         dst = GET_FIRST_BIT_SQUARE(single_push_target);
         src = dst - 8 + (16 * turn);
         if ((dst <= h8 && dst >= a8) || (dst >= a1 && dst <= h1)) {
-            move_list->list[i++] = CREATE_MOVE(src, dst, piece, Q, NO_CAPTURE, 0, 0, 0);
-            move_list->list[i++] = CREATE_MOVE(src, dst, piece, R, NO_CAPTURE, 0, 0, 0);
-            move_list->list[i++] = CREATE_MOVE(src, dst, piece, B, NO_CAPTURE, 0, 0, 0);
-            move_list->list[i++] = CREATE_MOVE(src, dst, piece, N, NO_CAPTURE, 0, 0, 0);
+            move_list->move[i++] = CREATE_MOVE(src, dst, piece, Q, NO_CAPTURE, 0, 0, 0);
+            move_list->move[i++] = CREATE_MOVE(src, dst, piece, R, NO_CAPTURE, 0, 0, 0);
+            move_list->move[i++] = CREATE_MOVE(src, dst, piece, B, NO_CAPTURE, 0, 0, 0);
+            move_list->move[i++] = CREATE_MOVE(src, dst, piece, N, NO_CAPTURE, 0, 0, 0);
         }
         else {
-            move_list->list[i++] = CREATE_MOVE(src, dst, piece, NO_PROMOTION, NO_CAPTURE, 0, 0, 0);
+            move_list->move[i++] = CREATE_MOVE(src, dst, piece, NO_PROMOTION, NO_CAPTURE, 0, 0, 0);
         }
         CLEAR_LSB(single_push_target);
     }
@@ -57,7 +57,7 @@ void generate_pawn_moves(const GameState *pos, int turn, int piece_offset, MoveL
             U64 epAttacks = pawn_attacks[turn][src] & (1ULL << pos->enpassant_square);
             if (epAttacks) {
                 dst = pos->enpassant_square;
-                move_list->list[i++] = CREATE_MOVE(src, dst, piece, NO_PROMOTION, p - piece_offset, 0, 1, 0);
+                move_list->move[i++] = CREATE_MOVE(src, dst, piece, NO_PROMOTION, p - piece_offset, 0, 1, 0);
             }
         }
 
@@ -67,13 +67,13 @@ void generate_pawn_moves(const GameState *pos, int turn, int piece_offset, MoveL
 
             // Promotion
             if ((dst <= h8 && dst >= a8) || (dst >= a1 && dst <= h1)) {
-                move_list->list[i++] = CREATE_MOVE(src, dst, piece, Q, victim, 0, 0, 0);
-                move_list->list[i++] = CREATE_MOVE(src, dst, piece, R, victim, 0, 0, 0);
-                move_list->list[i++] = CREATE_MOVE(src, dst, piece, B, victim, 0, 0, 0);
-                move_list->list[i++] = CREATE_MOVE(src, dst, piece, N, victim, 0, 0, 0);
+                move_list->move[i++] = CREATE_MOVE(src, dst, piece, Q, victim, 0, 0, 0);
+                move_list->move[i++] = CREATE_MOVE(src, dst, piece, R, victim, 0, 0, 0);
+                move_list->move[i++] = CREATE_MOVE(src, dst, piece, B, victim, 0, 0, 0);
+                move_list->move[i++] = CREATE_MOVE(src, dst, piece, N, victim, 0, 0, 0);
             }
             else {
-                move_list->list[i++] = CREATE_MOVE(src, dst, piece, NO_PROMOTION, victim, 0, 0, 0);
+                move_list->move[i++] = CREATE_MOVE(src, dst, piece, NO_PROMOTION, victim, 0, 0, 0);
             }
             CLEAR_LSB(piece_attacks);
         }
@@ -95,26 +95,26 @@ void generate_king_moves(const GameState *pos, int turn, int piece_offset, MoveL
     if (turn == WHITE) {
         if (castling_rights & WHITE_OO && !GET_SQUARE(occupancy, f1) && !GET_SQUARE(occupancy, g1)) {
             if (!(is_square_attacked(pos, e1, BLACK) || is_square_attacked(pos, f1, BLACK))) {
-                move_list->list[i++] = CREATE_MOVE(src, src + 2, K, NO_PROMOTION, NO_CAPTURE, 0, 0, 1);
+                move_list->move[i++] = CREATE_MOVE(src, src + 2, K, NO_PROMOTION, NO_CAPTURE, 0, 0, 1);
             }
         }
         if (castling_rights & WHITE_OOO && !GET_SQUARE(occupancy, b1) && !GET_SQUARE(occupancy, c1) && !
             GET_SQUARE(occupancy, d1)) {
             if (!(is_square_attacked(pos, e1, BLACK) || is_square_attacked(pos, d1, BLACK))) {
-                move_list->list[i++] = CREATE_MOVE(src, src - 2, K, NO_PROMOTION, NO_CAPTURE, 0, 0, 1);
+                move_list->move[i++] = CREATE_MOVE(src, src - 2, K, NO_PROMOTION, NO_CAPTURE, 0, 0, 1);
             }
         }
     }
     else {
         if (castling_rights & BLACK_OO && !GET_SQUARE(occupancy, f8) && !GET_SQUARE(occupancy, g8)) {
             if (!(is_square_attacked(pos, e8, WHITE) || is_square_attacked(pos, f8, WHITE))) {
-                move_list->list[i++] = CREATE_MOVE(src, src + 2, k, NO_PROMOTION, NO_CAPTURE, 0, 0, 1);
+                move_list->move[i++] = CREATE_MOVE(src, src + 2, k, NO_PROMOTION, NO_CAPTURE, 0, 0, 1);
             }
         }
         if (castling_rights & BLACK_OOO && !GET_SQUARE(occupancy, b8) && !GET_SQUARE(occupancy, c8) && !
             GET_SQUARE(occupancy, d8)) {
             if (!(is_square_attacked(pos, e8, WHITE) || is_square_attacked(pos, d8, WHITE))) {
-                move_list->list[i++] = CREATE_MOVE(src, src - 2, k, NO_PROMOTION, NO_CAPTURE, 0, 0, 1);
+                move_list->move[i++] = CREATE_MOVE(src, src - 2, k, NO_PROMOTION, NO_CAPTURE, 0, 0, 1);
             }
         }
     }
@@ -126,7 +126,7 @@ void generate_king_moves(const GameState *pos, int turn, int piece_offset, MoveL
         while (piece_attacks) {
             int dst = GET_FIRST_BIT_SQUARE(piece_attacks);
             int victim = (pos->mailbox[dst] == NO_PIECE) ? NO_CAPTURE : pos->mailbox[dst];
-            move_list->list[i++] = CREATE_MOVE(src, dst, piece, NO_PROMOTION, victim, 0, 0, 0);
+            move_list->move[i++] = CREATE_MOVE(src, dst, piece, NO_PROMOTION, victim, 0, 0, 0);
             CLEAR_LSB(piece_attacks);
         }
         CLEAR_LSB(piece_bb);
@@ -149,7 +149,7 @@ void generate_knight_moves(const GameState *pos, int turn, int piece_offset, Mov
         while (piece_attacks) {
             int dst = GET_FIRST_BIT_SQUARE(piece_attacks);
             int victim = (pos->mailbox[dst] == NO_PIECE) ? NO_CAPTURE : pos->mailbox[dst];
-            move_list->list[i++] = CREATE_MOVE(src, dst, piece, NO_PROMOTION, victim, 0, 0, 0);
+            move_list->move[i++] = CREATE_MOVE(src, dst, piece, NO_PROMOTION, victim, 0, 0, 0);
             CLEAR_LSB(piece_attacks);
         }
         CLEAR_LSB(piece_bb);
@@ -173,7 +173,7 @@ void generate_bishop_moves(const GameState *pos, int turn, int piece_offset, Mov
         while (piece_attacks) {
             int dst = GET_FIRST_BIT_SQUARE(piece_attacks);
             int victim = (pos->mailbox[dst] == NO_PIECE) ? NO_CAPTURE : pos->mailbox[dst];
-            move_list->list[i++] = CREATE_MOVE(src, dst, piece, NO_PROMOTION, victim, 0, 0, 0);
+            move_list->move[i++] = CREATE_MOVE(src, dst, piece, NO_PROMOTION, victim, 0, 0, 0);
             CLEAR_LSB(piece_attacks);
         }
         CLEAR_LSB(piece_bb);
@@ -197,7 +197,7 @@ void generate_rook_moves(const GameState *pos, int turn, int piece_offset, MoveL
         while (piece_attacks) {
             int dst = GET_FIRST_BIT_SQUARE(piece_attacks);
             int victim = (pos->mailbox[dst] == NO_PIECE) ? NO_CAPTURE : pos->mailbox[dst];
-            move_list->list[i++] = CREATE_MOVE(src, dst, piece, NO_PROMOTION, victim, 0, 0, 0);
+            move_list->move[i++] = CREATE_MOVE(src, dst, piece, NO_PROMOTION, victim, 0, 0, 0);
             CLEAR_LSB(piece_attacks);
         }
         CLEAR_LSB(piece_bb);
@@ -221,7 +221,7 @@ void generate_queen_moves(const GameState *pos, int turn, int piece_offset, Move
         while (piece_attacks) {
             int dst = GET_FIRST_BIT_SQUARE(piece_attacks);
             int victim = (pos->mailbox[dst] == NO_PIECE) ? NO_CAPTURE : pos->mailbox[dst];
-            move_list->list[i++] = CREATE_MOVE(src, dst, piece, NO_PROMOTION, victim, 0, 0, 0);
+            move_list->move[i++] = CREATE_MOVE(src, dst, piece, NO_PROMOTION, victim, 0, 0, 0);
             CLEAR_LSB(piece_attacks);
         }
         CLEAR_LSB(piece_bb);
@@ -233,7 +233,7 @@ MoveList generate_moves(const GameState *pos, int *size)
 {
     MoveList move_list;
     move_list.next_open = 0;
-    memset(move_list.list, 0, sizeof(Move) * MAX_MOVES);
+    memset(move_list.move, 0, sizeof(Move) * MAX_MOVES);
     int turn = pos->turn;
     int piece_offset = 6 * turn;
