+++ movegen.h
@@ -1,7 +1,7 @@
 #ifndef MOVEGEN_H
 #define MOVEGEN_H
 
-#include "list.h"
+#include "movelist.h"
 
 MoveList generate_moves(const GameState *position, int *size);
