+++ see.h
@@ -1,0 +1,7 @@
+#ifndef SEE_H
+#define SEE_H
+#include "position.h"
+
+int see(const GameState *pos, int square);
+
+#endif
